refactor(interfaces): move api to new layout
